feat: Raise error if wrong vLLM parameters are set
